+++ lib/src/eval/shared/stdlib/async.dart
@@ -2,6 +2,7 @@ import 'package:dart_eval/dart_eval.dart';
 import 'package:dart_eval/dart_eval_bridge.dart';
 import 'package:dart_eval/src/eval/shared/stdlib/async/stream.dart';
 import 'package:dart_eval/src/eval/shared/stdlib/async/stream_controller.dart';
+import 'package:dart_eval/src/eval/shared/stdlib/async/zone.dart';
 import 'async/future.dart';
 
 /// [EvalPlugin] for the `dart:async` library
@@ -16,30 +17,22 @@ class DartAsyncPlugin implements EvalPlugin {
     registry.defineBridgeClass($StreamSink.$declaration);
     registry.defineBridgeClass($Stream.$declaration);
     registry.defineBridgeClass($StreamController.$declaration);
+    registry.defineBridgeClass($Zone.$declaration);
   }
 
   @override
   void configureForRuntime(Runtime runtime) {
-    // Register Completer
     runtime.registerBridgeFunc(
         'dart:async', 'Completer.', const $Completer_new());
-
-    // Register StreamSubscription
-
-    // Register StreamSink
-
-    // Register Stream
     runtime.registerBridgeFunc('dart:async', 'Stream.empty', $Stream.$empty);
     runtime.registerBridgeFunc('dart:async', 'Stream.value', $Stream.$_value);
-    //runtime.registerBridgeFunc('dart:async', 'Stream.error', $Stream.$error);
-    //runtime.registerBridgeFunc('dart:async', 'Stream.fromFuture', $Stream.$fromFuture);
     runtime.registerBridgeFunc(
         'dart:async', 'Stream.fromIterable', $Stream.$fromIterable);
     runtime.registerBridgeFunc(
         'dart:async', 'Stream.periodic', $Stream.$periodic);
-
-    // Register StreamController
     runtime.registerBridgeFunc(
         'dart:async', 'StreamController.', $StreamController.$new);
+    runtime.registerBridgeFunc('dart:async', 'Zone.current*g', $Zone.$current);
+    runtime.registerBridgeFunc('dart:async', 'Zone.root*g', $Zone.$root);
   }
 }