+++ lib/src/eval/shared/stdlib/core.dart
@@ -14,6 +14,7 @@ import 'package:dart_eval/src/eval/shared/stdlib/core/object.dart';
 import 'package:dart_eval/src/eval/shared/stdlib/core/pattern.dart';
 import 'package:dart_eval/src/eval/shared/stdlib/core/regexp.dart';
 import 'package:dart_eval/src/eval/shared/stdlib/core/string_buffer.dart';
+import 'package:dart_eval/src/eval/shared/stdlib/core/symbol.dart';
 import 'package:dart_eval/src/eval/shared/stdlib/core/type.dart';
 import 'package:dart_eval/src/eval/shared/stdlib/core/uri.dart';
 import 'core/duration.dart';
@@ -38,6 +39,7 @@ class DartCorePlugin implements EvalPlugin {
     registry.defineBridgeClass($enumDeclaration);
     registry.defineBridgeClass($bool.$declaration);
     registry.defineBridgeClass($Function.$declaration);
+    registry.defineBridgeClass($Symbol.$declaration);
     registry.defineBridgeClass($num.$declaration);
     registry.defineBridgeClass($int.$declaration);
     registry.defineBridgeClass($double.$declaration);
@@ -91,5 +93,6 @@ class DartCorePlugin implements EvalPlugin {
         $RangeError.$checkValidRange);
     runtime.registerBridgeFunc('dart:core', 'RangeError.checkNotNegative',
         $RangeError.$checkNotNegative);
+    runtime.registerBridgeFunc('dart:core', 'Symbol.', $Symbol.$new);
   }
 }